src/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart

pydantic>=1.10
pydantic[email]

numpy
pandas
scikit-learn
sentence-transformers

torch
accelerate
transformers>=4.40.0
peft>=0.11.1
tokenizers

prometheus-client==0.20.0

requests

bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]